Added wheel_list Dictionary to class mf_vehicle_body which will keep track of the wheels and their position in the vector.
This way we can call the node name have it pick the correct wheel in the
array.

I'm sure I can come up with a better way to handle this and I am sure I
will but this will work for now.

We're just trying to get things working and then we can optimize and
start expanding even further.